move datatype check and label validation into valdiators directory
