+++ bids-validator/src/validators/bids.ts
@@ -3,14 +3,19 @@ import { FileTree } from '../files/filetree.ts'
 import { walkFileTree } from '../schema/walk.ts'
 import { loadSchema } from '../setup/loadSchema.ts'
 import { applyRules } from '../schema/applyRules.ts'
+import { isTopLevel, checkDatatypes, checkLabelFormat } from './filenames.ts'
 
 /**
  * Full BIDS schema validation entrypoint
  */
 export async function validate(fileTree: FileTree): Promise<> {
   const issues = []
-  const schemaDefs = await loadSchema()
+  const schema = await loadSchema()
   for await (const context of walkFileTree(fileTree)) {
-    applyRules(schemaDefs, context)
+    if (!isTopLevel(schema, context)) {
+      checkDatatypes(schema, context)
+      checkLabelFormat(schema, context)
+    }
+    applyRules(schema, context)
   }
 }
